fix: address Copilot review on HM dtype and Macenko docs
Cast histogram-matching float inputs to float32 before the CUDA
kernel, and align changelog/docs with the on-GPU Macenko path.

+++ src/stainx_cuda_torch/csrc/histogram_matching.cu
@@ -60,8 +60,9 @@ torch::Tensor histogram_matching_cuda(torch::Tensor input_images, torch::Tensor 
         was_uint8_or_high_range = true;
     } else {
         // Float is assumed [0, 1] — scale to uint8 (no max()>1 heuristic / sync).
+        // Cast to float32 before data_ptr<float>() so float16/float64 (e.g. AMP) are safe.
         needs_scale_back         = true;
-        torch::Tensor input_flat = input_images.contiguous().view(-1);
+        torch::Tensor input_flat = input_images.to(torch::kFloat32).contiguous().view(-1);
         images_uint8             = torch::empty({N, C, H, W}, torch::TensorOptions().dtype(torch::kUInt8).device(input_images.device()));
         int total_pixels         = N * C * H * W;
         int num_blocks           = (total_pixels + num_threads - 1) / num_threads;
